+++ internal/database/test/user_link_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/jmoiron/sqlx"
-	_ "github.com/mattn/go-sqlite3"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
@@ -12,7 +11,7 @@ import (
 )
 
 func setupUserLinkTestDB(t *testing.T) *sqlx.DB {
-	db, err := sqlx.Connect("sqlite3", ":memory:")
+	db, err := sqlx.Connect(database.DriverName, database.MemoryDSN(true))
 	require.NoError(t, err)
 	database.CreateTables(db)
 
@@ -28,8 +27,8 @@ func setupUserLinkTestDB(t *testing.T) *sqlx.DB {
 	require.NoError(t, err)
 
 	lst := &database.Lst{
-		Id:      1,
-		Name:    "TestList",
+		Id:          1,
+		Name:        "TestList",
 		OwnerUserId: 100,
 	}
 	err = database.CreateLst(db, lst)